# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Naroom API
  description: Naroom 모바일 앱을 위한 백엔드 API
  version: v1
servers:
- url: /
paths:
...
  /api/v1/auth/restore:
    post:
      tags:
      - auth-controller
      operationId: restore
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KakaoLoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSocialLoginResponse"
  /api/v1/auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRefreshResponse"
...
  /api/v1/auth/kakao/login:
    post:
      tags:
      - auth-controller
      operationId: kakaoLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KakaoLoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSocialLoginResponse"
  /api/v1/auth/google/login:
    post:
      tags:
      - auth-controller
      operationId: googleLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSocialLoginResponse"
  /api/v1/auth/apple/login:
    post:
      tags:
      - auth-controller
      operationId: appleLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleLoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSocialLoginResponse"
...
components:
  schemas:
...
    DeviceInfo:
      type: object
      properties:
        installationKey:
          type: string
        platform:
          type: string
        appVersion:
          type: string
    KakaoLoginRequest:
      type: object
      properties:
        providerAccessToken:
          type: string
          minLength: 1
        device:
          $ref: "#/components/schemas/DeviceInfo"
      required:
      - device
      - providerAccessToken
    AccountSummary:
      type: object
      properties:
        memberId:
          type: string
          format: uuid
        displayName:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - LOCKED
          - PENDING_DELETION
        onboardingCompletedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    ApiResponseSocialLoginResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/SocialLoginResponse"
    SessionSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
    SocialLoginResponse:
      type: object
      properties:
        tokenType:
          type: string
        accessToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        refreshToken:
          type: string
        refreshTokenExpiresAt:
          type: string
          format: date-time
        session:
          $ref: "#/components/schemas/SessionSummary"
        account:
          $ref: "#/components/schemas/AccountSummary"
        nextAction:
          type: string
          enum:
          - COMPLETE_ONBOARDING
          - ENTER_APP
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
        installationKey:
          type: string
      required:
      - refreshToken
    ApiResponseRefreshResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/RefreshResponse"
    RefreshResponse:
      type: object
      properties:
        tokenType:
          type: string
        accessToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        refreshToken:
          type: string
        refreshTokenExpiresAt:
          type: string
          format: date-time
        session:
          $ref: "#/components/schemas/SessionSummary"
    GoogleLoginRequest:
      type: object
      properties:
        idToken:
          type: string
          minLength: 1
        device:
          $ref: "#/components/schemas/DeviceInfo"
      required:
      - device
      - idToken
    AppleLoginRequest:
      type: object
      properties:
        identityToken:
          type: string
          minLength: 1
        rawNonce:
          type: string
          minLength: 1
        fullName:
          type: string
        device:
          $ref: "#/components/schemas/DeviceInfo"
      required:
      - device
      - identityToken
      - rawNonce
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT